Add Textures class and new folder structure
