+++ aicargame/main.py
@@ -1,6 +1,7 @@
 import pygame
 from pygame import Vector2
-from drawableobject import Player
+
+from aicargame.objects.player import Player
 
 pygame.init()
 
@@ -8,7 +9,7 @@ win = pygame.display.set_mode((500, 500))
 
 pygame.display.set_caption("AICarGame")
 
-obj = Player(Vector2(50, 50), Vector2(100, 100), "ball.gif")
+obj = Player(Vector2(50, 50), Vector2(100, 100))
 
 all_sprites = pygame.sprite.Group()
 all_sprites.add(obj)
