+++ aicargame/objects/player.py
@@ -1,34 +1,14 @@
 import pygame
 from pygame import Vector2
 
-
-class DrawableObject(pygame.sprite.Sprite):
-    def __init__(self, position: Vector2, size: Vector2, texture=None, color=None):
-        if texture is None and color is None:
-            raise NotImplementedError
-
-        super().__init__()
-
-        if texture is None:
-            self.image = pygame.Surface(size)
-            self.image.fill(color)
-        else:
-            try:
-                texture = pygame.image.load(texture)
-            except Exception as ex:
-                self.image.fill((255, 255, 255))
-                print(str(ex))
-            else:
-                texture = pygame.transform.scale(texture, size)
-                self.image = texture
-
-        self.rect = self.image.get_rect()
-        self.rect.center = position
+from aicargame.objects.drawableobject import DrawableObject
+from aicargame.textures.textures import Textures
 
 
 class Player(DrawableObject):
-    def __init__(self, position: Vector2, size: Vector2, texture):
-        super().__init__(position, size, texture)
+    def __init__(self, position: Vector2, size: Vector2):
+        super().__init__(position, size, Textures.PLAYER)
+        print(str(Textures.PLAYER))
         self.__velocity = 10
         self.__x_border = 10
         self.__y_border = 10
